dev-requirements.txt
-r requirements.txt
exam==0.10.5
ipdb==0.8
ipython==2.4.1
pytest-django==2.8.0
pytest==2.6.4
